Allow overriding AsicContainerParser::parseSignatures
Opening an existing container with signatures in the current
version will trigger loading each file fully in memory. If any
of the files does not fit, then the loader will crash with OOM.

Allow overriding parseSignatures so that custom implementations
can implement workarounds such as using DigestDocument for signature
validation.

Signed-off-by: Märt Bakhoff <mbakhoff@sigil.red>